requirements.txt
google-generativeai
psycopg2-binary
python-dotenv
pgvector 
# Add other specific libraries if used, e.g., sentence-transformers if not using dummy embeddings
